[RED] Create interface for: create planned payment
